全部功能测试完成
2020年3月23日20:41:15

+++ can/src/main/java/huat/wubeibei/smartscreenserver/modal/Signal.java
@@ -6,7 +6,7 @@ public class Signal {
     private int SignalLength;
     private int StartBytePosition;
     private int StartBitPosition;
-    private int Resolution;
+    private double Resolution;
     private int Offset;
     private int MinValue;
     private int MaxValue;
@@ -57,11 +57,11 @@ public class Signal {
         StartBitPosition = startBitPosition;
     }
 
-    public int getResolution() {
+    public double getResolution() {
         return Resolution;
     }
 
-    public void setResolution(int resolution) {
+    public void setResolution(double resolution) {
         Resolution = resolution;
     }
